+++ packages/transaction/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @abloatai/transaction
 
+## 0.66.2
+
 ## 0.66.1
 
 ## 0.66.0